Charms + Break Vue into component files
